pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "earth2-physicsnemo-weather-recipes"
version = "0.1.0"
description = "Lightweight Earth-2 / PhysicsNeMo-aligned weather ML recipe starter."
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [
  { name = "Catherine Ravenscroft" }
]
dependencies = [
  "matplotlib>=3.8",
  "numpy>=1.26",
  "PyYAML>=6.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0",
  "ruff>=0.4",
]

[tool.setuptools]
package-dir = { "" = "src" }

[tool.setuptools.packages.find]
where = ["src"]
